move testing and benchmarking stuff into a folder
a baby library
