+++ core/incremental/ddTests.ts
@@ -4,7 +4,7 @@ import { Rec, Statement } from "../types";
 import { formatOutput, IncrementalInterpreter } from "./interpreter";
 import { getJoinInfo } from "./build";
 import { fsLoader } from "../fsLoader";
-import { Suite } from "../../util/testing";
+import { Suite } from "../../util/testBench/testing";
 import { ProcessFn, runDDTestAtPath, TestOutput } from "../../util/ddTest";
 import { graphvizOut, jsonOut } from "../../util/ddTest/types";
 import { prettyPrintGraph } from "../../util/graphviz";
